Update to guard against potential NPE (#2640)
Also some grammar fixes

+++ src/Agent.Sdk/ProcessInvoker.MacLinux.cs
@@ -13,21 +13,27 @@ namespace Microsoft.VisualStudio.Services.Agent.Util
     {
         private async Task<bool> SendPosixSignal(PosixSignals signal, TimeSpan timeout)
         {
+            if (_proc == null)
+            {
+                Trace.Info($"Process already exited, no need to send {signal}.");
+                return true;
+            }
+
             Trace.Info($"Sending {signal} to process {_proc.Id}.");
             int errorCode = kill(_proc.Id, (int)signal);
             if (errorCode != 0)
             {
-                Trace.Info($"{signal} signal doesn't fire successfully.");
+                Trace.Info($"{signal} signal did not fire successfully.");
                 Trace.Info($"Error code: {errorCode}.");
                 return false;
             }
 
-            Trace.Info($"Successfully send {signal} to process {_proc.Id}.");
+            Trace.Info($"Successfully sent {signal} to process {_proc.Id}.");
             Trace.Info($"Waiting for process exit or {timeout.TotalSeconds} seconds after {signal} signal fired.");
             var completedTask = await Task.WhenAny(Task.Delay(timeout), _processExitedCompletionSource.Task);
             if (completedTask == _processExitedCompletionSource.Task)
             {
-                Trace.Info("Process exit successfully.");
+                Trace.Info("Process exited successfully.");
                 return true;
             }
             else
